refactor(frontend): decompose 957-line Sidebar.tsx into focused modules
Extract four sub-modules so each file has a single responsibility:

- sidebar-utils.ts: isActive, filterNavItems, groupNavItems, NavGroup
  types, and shared sizing constants (COLLAPSED_NAV_ITEM_SIZE, etc.)
- NavItem.tsx: page navigation item (link + active highlight + tooltip)
- NavLinkItem.tsx: external footer link item
- NavSection.tsx: collapsible section header + item list

Sidebar.tsx is now 580 lines (was 957), acting as the orchestrator
that assembles the brand header, org/user menus, and nav groups.
